Updated Main

+++ constructSpringApp1/src/main/java/com/masai/model/Planter.java
@@ -38,7 +38,7 @@ public class Planter {
 	private Integer planterStock;
 	@NotNull(message = "Planter Cost should not be null")
 	@Min(value = 1, message = "Planter cost is must be greater than equal to 1")
-	private Integer planterCost;
+	private Double planterCost;
 
 	@OneToOne(cascade = CascadeType.ALL)
 	private Seed seed;
